renamed use cases directories to 'log_in' and 'log out'
