--- src/main/java/LogOut/interface_adapters/LogOutControllerData.java
@@ -1,14 +1,0 @@
-package LogOut.interface_adapters;
-
-// data to be processed
-public class LogOutControllerData {
-    private final String logOutRequest;
-
-    public LogOutControllerData(String logOutRequest) {
-        this.logOutRequest = logOutRequest;
-    }
-
-    public String getLogOutRequest(){
-        return this.logOutRequest;
-    }
-}
