--- src/main/java/LogOut/use_case/LogOutInputBoundary.java
@@ -1,10 +1,0 @@
-package LogOut.use_case;
-
-public interface LogOutInputBoundary {
-
-    /**
-     * Interface which contains a logout request
-     * @param requestModel a request to logout made by the user
-     */
-    void logOut(LogOutRequestModel requestModel);
-}
