--- src/main/java/LogOut/use_case/LogOutOutputBoundary.java
@@ -1,9 +1,0 @@
-package LogOut.use_case;
-
-/**
- * this interface indicates the logout was successful
- */
-public interface LogOutOutputBoundary {
-
-    void present(LogOutResponseModel responseModel);
-}
