--- src/main/java/LogOut/use_case/LogOutResponseModel.java
@@ -1,19 +1,0 @@
-package LogOut.use_case;
-
-public class LogOutResponseModel {
-
-
-    private final boolean success;
-
-    /**
-     *
-     * @param success Parameter indicating the logout request was successful
-     */
-    public LogOutResponseModel(boolean success){
-        this.success = success;
-    }
-
-    public boolean isSuccess(){
-        return this.success;
-    }
-}
