--- src/main/java/logIn/use_case/LogInOutputBoundary.java
@@ -1,8 +1,0 @@
-package logIn.use_case;
-
-/**
- * This Interface contains the result of the log in
- */
-public interface LogInOutputBoundary {
-    void present(LogInResponseModel responseModel);
-}
