--- src/main/java/logIn/use_case/exceptions/UserException.java
@@ -1,14 +1,0 @@
-package logIn.use_case.exceptions;
-
-public class UserException extends Exception{
-    private final String error;
-
-    public UserException(String error){
-        this.error = error;
-    }
-
-    @Override
-    public String getMessage(){
-        return error;
-    }
-}
