+++ src/main/java/log_in/use_case/LogInOutputBoundary.java
@@ -1,0 +1,8 @@
+package log_in.use_case;
+
+/**
+ * This Interface contains the result of the log in
+ */
+public interface LogInOutputBoundary {
+    void present(LogInResponseModel responseModel);
+}
