+++ src/main/java/log_out/use_case/LogOutOutputBoundary.java
@@ -1,0 +1,9 @@
+package log_out.use_case;
+
+/**
+ * this interface indicates the logout was successful
+ */
+public interface LogOutOutputBoundary {
+
+    void present(LogOutResponseModel responseModel);
+}
