+++ src/main/java/log_out/use_case/LogOutRequestModel.java
@@ -1,0 +1,19 @@
+package log_out.use_case;
+
+public class LogOutRequestModel {
+
+    private final String logOutRequest;
+
+
+    /**
+     * Initialize a log_out Request
+     * @param logOutRequest A string to indicate a logout request has been made
+     */
+    public LogOutRequestModel(String logOutRequest){
+        this.logOutRequest  = logOutRequest;
+    }
+
+    public String getLogOutRequest(){
+        return this.logOutRequest;
+    }
+}
