+++ src/main/java/log_out/use_case/LogOutResponseModel.java
@@ -1,0 +1,19 @@
+package log_out.use_case;
+
+public class LogOutResponseModel {
+
+
+    private final boolean success;
+
+    /**
+     *
+     * @param success Parameter indicating the logout request was successful
+     */
+    public LogOutResponseModel(boolean success){
+        this.success = success;
+    }
+
+    public boolean isSuccess(){
+        return this.success;
+    }
+}
